- **Unit Tests**: Implemented comprehensive unit tests for core services and repositories, including `PassengerService`, `UserService`, `TokenService`, and utility classes like `JwtUtility` and `PasswordHasher`. Addressed errors related to `IEnumerable` indexing and method arguments within tests. - **Repository Unit Tests**: Added unit tests for `UserRepository`, `PassengerRepository`, and `TokenRepository`, improving test coverage and ensuring data integrity. - **Controller & Repository Integration Tests**: Implemented integration tests for controllers and repositories, focusing on real data interaction via MongoDB. - **Helpers & Events Tests**: Developed initial unit tests for helper methods and event handlers. - **Dummy Tests**: Some unit and integration tests are placeholders (marked with `[Skip]`), allowing the project to compile and run while tests are still being fully implemented. These will be addressed in upcoming commits.
- **CI/CD Pipeline**: Enhanced the CI/CD pipeline, adding manual approval steps for deployment to production, and refined the pipeline for seamless Docker and Kubernetes deployment.
- **Kubernetes Configurations**: Updated Kubernetes YAML files for both API and MongoDB deployments, including `LoadBalancer` services, health checks, and environment configurations.
- **Logging & Monitoring**: Introduced logging for token cleanup tasks and fixed return types within repositories. Improved logging setup to provide better traceability during production operations.
- **Docker**: Improved Docker configuration for local and production environments. Separated MongoDB seed image generation from the main API image to enhance modularity and reduce complexity.

- **JWT Authentication**: Refactored JWT authentication logic and made enhancements to the Swagger UI configuration to support JWT Bearer token authentication. Updated documentation to explain that tokens must be prefixed with "Bearer" for successful authorization.
- **Swagger UI**: Configured Swagger to expose API endpoints and integrated JWT-based authentication for API documentation and testing.

- **Database Seeding**: Refactored the MongoDB seeding process to ensure proper population of the `Passengers` collection from the `titanic.csv` file, both in local and Kubernetes deployments.

- **Codebase Refactoring**: Refactored various parts of the codebase, including simplifying database connection logic by using `DatabaseConfig` and environment variables. Cleaned up unused and redundant code.
- **Health Checks**: Integrated health check endpoints (`/health/live` and `/health/ready`) for Kubernetes readiness and liveness probes.

- **.env Handling**: Removed the `.env` file from the repository and added it to `.gitignore` to prevent accidental commits of sensitive information, such as the Docker Personal Access Token. Cleaned up multiple commits related to `.env` file handling.
- **Refactoring and Cleanup**: Performed significant refactoring and removed unnecessary and redundant commits to maintain a cleaner code history and improve maintainability.
- **Reverted Changes**: Addressed errors caused by incorrect `.env` handling by reverting several commits and reapplying the proper changes.

- **Dummy Tests**: The current placeholder tests will be completed in upcoming commits, ensuring full test coverage across all services and repositories.

+++ Poseidon/DTOs/PassengerDTO.cs
@@ -7,7 +7,7 @@
         public int? Pclass { get; set; }
         public string Name { get; set; }
         public string Sex { get; set; }
-        public int? Age { get; set; }
+        public double? Age { get; set; }
         public int SiblingsOrSpousesAboard { get; set; }
         public int ParentsOrChildrenAboard { get; set; }
         public double? Fare { get; set; }
