+++ Poseidon/Interfaces/IRepositories/IPassengerRepository.cs
@@ -7,8 +7,8 @@ namespace Poseidon.Interfaces.IRepositories
         // Additional passenger-specific methods if needed
         Task<IEnumerable<Passenger>> GetByClassAsync(int classNumber);
         Task<IEnumerable<Passenger>> GetByGenderAsync(string gender);
-        Task<IEnumerable<Passenger>> GetByAgeRangeAsync(int minAge, int maxAge);
-        Task<IEnumerable<Passenger>> GetByFareRangeAsync(decimal minFare, decimal maxFare);
+        Task<IEnumerable<Passenger>> GetByAgeRangeAsync(double minAge, double maxAge);
+        Task<IEnumerable<Passenger>> GetByFareRangeAsync(double minFare, double maxFare);
         Task<IEnumerable<Passenger>> GetSurvivorsAsync();
         Task<double> GetSurvivalRateAsync();
     }
